requirement.txt

fastapi==0.111.0
uvicorn[standard]==0.23.2

pinecone-client==2.2.0


transformers==4.34.0
sentence-transformers==2.2.2
torch==2.2.0


cohere==5.9.1


python-dotenv==1.0.1
tiktoken==0.4.0

python-multipart